refactor(frontend): group detail page shells under pages/record/detail/
Move the single-record detail page-kind into pages/record/detail/: the shell
(RecordDetailShell, HeroHeader), the taxonomy detail layouts (Taxonomy/
Hierarchical*), and the detail-page sidebars/sections (ModelSpecsSidebar,
ModelHierarchy, CreditsList, ExternalLinksSidebarSection,
TaxonomyLinkSidebarSection), with co-located tests and fixtures.

All are multi-route page bodies. Kept flat: 20 files, in line with the sibling
pages/listing/ folder, with no clean sub-boundary to split on.
