+++ frontend/src/lib/components/pages/record/detail/CreditsList.svelte
