+++ frontend/src/lib/components/pages/record/detail/ExternalLinksSidebarSection.svelte
