+++ frontend/src/lib/components/pages/record/detail/HeroHeader.svelte
