+++ frontend/src/lib/components/pages/record/detail/HierarchicalTaxonomyChildrenAccordion.fixture.svelte
