+++ frontend/src/lib/components/pages/record/detail/HierarchicalTaxonomyChildrenAccordion.svelte
