+++ frontend/src/lib/components/pages/record/detail/HierarchicalTaxonomyMobileMetaBar.dom.test.ts
