+++ frontend/src/lib/components/pages/record/detail/HierarchicalTaxonomyMobileMetaBar.fixture.svelte
