+++ frontend/src/lib/components/pages/record/detail/HierarchicalTaxonomyMobileMetaBar.svelte
