+++ frontend/src/lib/components/pages/record/detail/HierarchicalTaxonomySidebar.dom.test.ts
