+++ frontend/src/lib/components/pages/record/detail/HierarchicalTaxonomySidebar.fixture.svelte
