+++ frontend/src/lib/components/pages/record/detail/ModelSpecsSidebar.svelte
