+++ frontend/src/lib/components/pages/record/detail/RecordDetailShell.svelte
