+++ frontend/src/lib/components/pages/record/detail/RecordDetailShell.test.ts
