+++ frontend/src/lib/components/pages/record/detail/SimpleTaxonomyDetailLayout.svelte
