+++ frontend/src/lib/components/pages/record/detail/TaxonomyLinkSidebarSection.svelte
