+++ frontend/src/routes/locations/[...path]/+layout.svelte
@@ -5,7 +5,7 @@
   import { auth } from '$lib/auth.svelte';
   import { WIDE_BREAKPOINT, pageTitle } from '$lib/constants';
   import PageActionBar from '$lib/components/layout/page/PageActionBar.svelte';
-  import RecordDetailShell from '$lib/components/RecordDetailShell.svelte';
+  import RecordDetailShell from '$lib/components/pages/record/detail/RecordDetailShell.svelte';
   import SectionEditorHost from '$lib/components/SectionEditorHost.svelte';
   import SidebarList from '$lib/components/layout/page/sidebar/SidebarList.svelte';
   import SidebarListItem from '$lib/components/layout/page/sidebar/SidebarListItem.svelte';
