+++ frontend/src/routes/manufacturers/[slug]/+layout.svelte
@@ -10,7 +10,7 @@
   import ExpandableSidebarList from '$lib/components/layout/page/sidebar/ExpandableSidebarList.svelte';
   import LocationLink from '$lib/components/LocationLink.svelte';
   import PageActionBar from '$lib/components/layout/page/PageActionBar.svelte';
-  import RecordDetailShell from '$lib/components/RecordDetailShell.svelte';
+  import RecordDetailShell from '$lib/components/pages/record/detail/RecordDetailShell.svelte';
   import SectionEditorHost from '$lib/components/SectionEditorHost.svelte';
   import SidebarList from '$lib/components/layout/page/sidebar/SidebarList.svelte';
   import SidebarListItem from '$lib/components/layout/page/sidebar/SidebarListItem.svelte';
