+++ frontend/src/routes/titles/[slug]/+layout.svelte
@@ -6,19 +6,19 @@
   import MetaTags from '$lib/components/layout/site/MetaTags.svelte';
   import { metaDescriptionFor } from '$lib/components/layout/site/meta-tags';
   import JsonLd from '$lib/components/layout/site/JsonLd.svelte';
-  import ExternalLinksSidebarSection from '$lib/components/ExternalLinksSidebarSection.svelte';
+  import ExternalLinksSidebarSection from '$lib/components/pages/record/detail/ExternalLinksSidebarSection.svelte';
   import { externalLinks } from '$lib/entities/external-links';
   import { model as modelInfo } from '$lib/entities/model';
   import { title as titleInfo } from '$lib/entities/title';
-  import ModelHierarchy from '$lib/components/ModelHierarchy.svelte';
-  import ModelSpecsSidebar from '$lib/components/ModelSpecsSidebar.svelte';
+  import ModelHierarchy from '$lib/components/pages/record/detail/ModelHierarchy.svelte';
+  import ModelSpecsSidebar from '$lib/components/pages/record/detail/ModelSpecsSidebar.svelte';
   import PageActionBar from '$lib/components/layout/page/PageActionBar.svelte';
-  import RecordDetailShell from '$lib/components/RecordDetailShell.svelte';
+  import RecordDetailShell from '$lib/components/pages/record/detail/RecordDetailShell.svelte';
   import SectionEditorHost from '$lib/components/SectionEditorHost.svelte';
   import SidebarList from '$lib/components/layout/page/sidebar/SidebarList.svelte';
   import SidebarListItem from '$lib/components/layout/page/sidebar/SidebarListItem.svelte';
   import SidebarSection from '$lib/components/layout/page/sidebar/SidebarSection.svelte';
-  import TaxonomyLinkSidebarSection from '$lib/components/TaxonomyLinkSidebarSection.svelte';
+  import TaxonomyLinkSidebarSection from '$lib/components/pages/record/detail/TaxonomyLinkSidebarSection.svelte';
   import NeedsReviewBanner from './_components/NeedsReviewBanner.svelte';
   import { getMenuItemAction, type EditSectionMenuItem } from '$lib/components/edit-section-menu';
   import { WIDE_BREAKPOINT } from '$lib/constants';
