+++ frontend/src/routes/titles/[slug]/+page.svelte
@@ -2,10 +2,10 @@
   import { resolve } from '$app/paths';
   import AccordionSection from '$lib/components/ui/AccordionSection.svelte';
   import CardGrid from '$lib/components/collections/grid/CardGrid.svelte';
-  import CreditsList from '$lib/components/CreditsList.svelte';
+  import CreditsList from '$lib/components/pages/record/detail/CreditsList.svelte';
   import MachineCard from '$lib/components/collections/cards/MachineCard.svelte';
   import MediaGrid from '$lib/components/media/MediaGrid.svelte';
-  import ModelSpecsSidebar from '$lib/components/ModelSpecsSidebar.svelte';
+  import ModelSpecsSidebar from '$lib/components/pages/record/detail/ModelSpecsSidebar.svelte';
   import RichTextOverviewAccordion from '$lib/components/markdown/RichTextOverviewAccordion.svelte';
   import RichTextReferencesAccordion from '$lib/components/markdown/RichTextReferencesAccordion.svelte';
   import { createRichTextAccordionState } from '$lib/components/markdown/rich-text-accordion-state.svelte';
